requirements.txt
aiohttp
discord.py==2.5.2
python-dotenv
requests
schedule
unidecode
typing-extensions
aiosqlite
aiofiles
types-aiofiles